+++ .isort.cfg
@@ -1,0 +1,3 @@
+[settings]
+profile=black
+line_length=88
