+++ requirements-dev.txt
@@ -1,0 +1,1 @@
+isort>=5.10.0
